+++ aes-kw/CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Added
+- `AssociatedOid` implementations ([#35])
+
+### Changed
+- Bump `aes` dependency to v0.9 ([#34])
+- `Kek` type is split into separate `AesKw` and `AesKwp` types ([#40])
+- `wrap` and `unwrap` methods now return resulting slice ([#40])
+
+### Removed
+- `Kek::new` inherent method in favor of implementing `InnerInit` ([#40])
+- `From`/`Into` impls from key into key-wrapper types ([#40])
+- `IV`, `KWP_IV_PREFIX`, and `KWP_MAX_LEN` constants ([#40])
+
+[#34]: https://github.com/RustCrypto/key-wraps/pull/34
+[#35]: https://github.com/RustCrypto/key-wraps/pull/35
+[#40]: https://github.com/RustCrypto/key-wraps/pull/40
+
 ## 0.2.1 (2022-04-20)
 ### Changed
 - Use `encrypt_with_backend`/`decrypt_with_backend` methods ([#19])